+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- When it is October, count down until 1st of November instead of the last day of October.
+
 ### Fixed
 - Fixed comment in code.
+- Fixed English.
+- Fixed CSS problems.
 
 ## [1.0.0] - 2019-08-26
 ### Added